warning removed

+++ tests/ODC_tests/TCPSocketManagerTests.cpp
@@ -204,7 +204,7 @@ TEST_CASE(SUITE("ManyStrings"))
 			{
 				count++;
 				auto ch = buf.sgetc();
-				if((count+recv_offset)%256 != ch)
+				if(static_cast<int>((count+recv_offset)%256) != ch)
 				{
 					recv_offset = ch - count%256;
 					odc::spdlog_get("opendatacon")->critical("{}: Possible out of order or lost data. Recieved {}, count {} ({}), offset {}",sock1 ? "Sock1" : "Sock2",(uint8_t)ch,count,count%256,recv_offset);
